refactor(cards): 카드 총합을 구하는 책임 이동
