+++ src/main/java/blackjack/domain/User.java
@@ -11,6 +11,6 @@ public class User extends Participant {
 
     @Override
     public boolean isDrawable() {
-        return calculateTotalValue() < STANDARD;
+        return cards.calculateTotalValue() < STANDARD;
     }
 }
